+++ src/main/java/com/sparta/deliveryi/user/infrastructure/keycloak/service/AuthRegister.java
@@ -1,6 +1,6 @@
 package com.sparta.deliveryi.user.infrastructure.keycloak.service;
 
-import com.sparta.deliveryi.user.infrastructure.keycloak.KeycloakUser;
+import com.sparta.deliveryi.user.infrastructure.keycloak.dto.KeycloakUser;
 import com.sparta.deliveryi.user.infrastructure.keycloak.dto.KeycloakRegisterRequest;
 import jakarta.validation.Valid;
