+++ src/main/java/com/sparta/deliveryi/user/infrastructure/keycloak/service/KeycloakQueryService.java
@@ -5,7 +5,7 @@ import com.sparta.deliveryi.user.domain.UserRole;
 import com.sparta.deliveryi.user.infrastructure.keycloak.KeycloakException;
 import com.sparta.deliveryi.user.infrastructure.keycloak.KeycloakMessageCode;
 import com.sparta.deliveryi.user.infrastructure.keycloak.KeycloakProperties;
-import com.sparta.deliveryi.user.infrastructure.keycloak.KeycloakUser;
+import com.sparta.deliveryi.user.infrastructure.keycloak.dto.KeycloakUser;
 import jakarta.ws.rs.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.keycloak.admin.client.Keycloak;
